chore: move test files into src directory
