+++ jest.config.cjs
@@ -14,7 +14,7 @@ module.exports = {
     "(.+)\\.js": "$1"
   },
   testMatch: [
-    '**/test/**/*.test.ts'
+    '**/src/**/*.test.ts'
   ],
 
   // Code Coverage
