+++ src/perform-backmerge.test.ts
@@ -1,7 +1,8 @@
-import {performBackmerge} from "../src/perform-backmerge";
-import Git from "../src/helpers/git";
+import {performBackmerge} from "./perform-backmerge";
+import Git from "./helpers/git";
 import {instance, mock, verify, when, anyString, anything} from "ts-mockito";
-import {resolveConfig} from "../src/helpers/resolve-config";
+import {resolveConfig} from "./helpers/resolve-config";
+import {Context} from "semantic-release";
 
 jest.mock('semantic-release/lib/get-git-auth-url.js', () => ({
     __esModule: true,
@@ -9,8 +10,8 @@ jest.mock('semantic-release/lib/get-git-auth-url.js', () => ({
 }));
 
 class NullLogger {
-    log(message) {}
-    error(message) {}
+    log(_message: string) {}
+    error(_message: string) {}
 }
 
 const realProcessExit = process.exit;
@@ -28,7 +29,7 @@ describe("perform-backmerge", () => {
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context;
         performBackmerge(instance(mockedGit), {backmergeBranches: ['develop']}, context)
             .then(() => {
                 verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
@@ -401,7 +402,7 @@ describe("perform-backmerge with deprecated branches setting", () => {
                 verify(mockedLogger.error('Invalid branch configuration found and ignored.')).never();
                 verify(mockedGit.checkout('master')).once();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.checkout('develop')).once();
                 verify(mockedGit.rebase('master')).once();
                 verify(mockedGit.push('my-repo', 'develop', false)).once();
@@ -420,14 +421,14 @@ describe("perform-backmerge with deprecated branches setting", () => {
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context;
         performBackmerge(instance(mockedGit), resolveConfig({}), context)
             .then(() => {
                 verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
                 verify(mockedLogger.error('Invalid branch configuration found and ignored.')).never();
                 verify(mockedGit.checkout('master')).once();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.checkout('develop')).once();
                 verify(mockedGit.rebase('master')).once();
                 verify(mockedGit.push('my-repo', 'develop', false)).once();
@@ -446,12 +447,12 @@ describe("perform-backmerge with deprecated branches setting", () => {
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context;
         performBackmerge(instance(mockedGit), {branches: ['master'], allowSameBranchMerge: true}, context)
             .then(() => {
                 verify(mockedLogger.log('Performing back-merge into develop branch "master".')).once();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.checkout('master')).once();
                 verify(mockedGit.rebase('master')).never();
                 verify(mockedGit.push('my-repo', 'master', false)).once();
@@ -463,7 +464,7 @@ describe("perform-backmerge with deprecated branches setting", () => {
     it("disallow merging into the same branch", async () => {
         const mockedGit = mock(Git);
         const mockedLogger = mock(NullLogger);
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context;
         await performBackmerge(instance(mockedGit), {branches: ['master'], allowSameBranchMerge: false}, context);
         verify(mockedLogger.error(anyString())).once();
     });
@@ -478,12 +479,12 @@ describe("perform-backmerge with deprecated branches setting", () => {
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context;
         performBackmerge(instance(mockedGit), {branches: ['${branch.name}'], allowSameBranchMerge: true}, context)
             .then(() => {
                 verify(mockedLogger.log('Performing back-merge into develop branch "master".')).once();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.checkout('master')).once();
                 verify(mockedGit.rebase('master')).never();
                 verify(mockedGit.push('my-repo', 'master', false)).once();
@@ -502,13 +503,13 @@ describe("perform-backmerge with deprecated branches setting", () => {
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context;
         performBackmerge(instance(mockedGit), {branches: ['develop']}, context)
             .then(() => {
                 verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
                 verify(mockedGit.checkout('master')).once();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.checkout('develop')).once();
                 verify(mockedGit.rebase('master')).once();
                 verify(mockedGit.push('my-repo', 'develop', false)).once();
@@ -527,13 +528,13 @@ describe("perform-backmerge with deprecated branches setting", () => {
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context;
         performBackmerge(instance(mockedGit), {branches: ['develop'], forcePush: true}, context)
             .then(() => {
                 verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
                 verify(mockedGit.checkout('master')).once();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.checkout('develop')).once();
                 verify(mockedGit.rebase('master')).once();
                 verify(mockedGit.push('my-repo', 'develop', true)).once();
@@ -554,7 +555,7 @@ describe("perform-backmerge with deprecated branches setting", () => {
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
 
         await performBackmerge(
             instance(mockedGit),
@@ -567,7 +568,7 @@ describe("perform-backmerge with deprecated branches setting", () => {
         verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
         verify(mockedGit.checkout('master')).once();
         verify(mockedGit.configFetchAllRemotes()).once();
-        verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+        verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
         verify(mockedGit.checkout('develop')).once();
         verify(mockedGit.rebase('master')).once();
         verify(mockedGit.commit('my-commit-message')).once();
@@ -586,7 +587,7 @@ describe("perform-backmerge with deprecated branches setting", () => {
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context;
 
         await performBackmerge(
             instance(mockedGit),
@@ -600,7 +601,7 @@ describe("perform-backmerge with deprecated branches setting", () => {
         verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
         verify(mockedGit.checkout('master')).once();
         verify(mockedGit.configFetchAllRemotes()).once();
-        verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+        verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
 
         const checkoutDevelopAction = mockedGit.checkout('develop');
         verify(mockedGit.stash()).calledBefore(checkoutDevelopAction);
@@ -624,7 +625,7 @@ describe("perform-backmerge with deprecated branches setting", () => {
         when(mockedGit.merge(anyString(), anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
 
         await performBackmerge(
             instance(mockedGit),
@@ -637,7 +638,7 @@ describe("perform-backmerge with deprecated branches setting", () => {
         verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
         verify(mockedGit.checkout('master')).once();
         verify(mockedGit.configFetchAllRemotes()).once();
-        verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+        verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
 
         verify(mockedGit.checkout('develop')).once();
         verify(mockedGit.merge('master', 'none')).once();
@@ -657,7 +658,7 @@ describe("perform-backmerge with deprecated branches setting", () => {
         when(mockedGit.merge(anyString(), anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context;
 
         await performBackmerge(
             instance(mockedGit),
@@ -671,7 +672,7 @@ describe("perform-backmerge with deprecated branches setting", () => {
         verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
         verify(mockedGit.checkout('master')).once();
         verify(mockedGit.configFetchAllRemotes()).once();
-        verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+        verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
 
         verify(mockedGit.checkout('develop')).once();
         verify(mockedGit.merge('master', 'ours')).once();
@@ -691,7 +692,7 @@ describe("perform-backmerge with deprecated branches setting", () => {
         when(mockedGit.merge(anyString(), anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
 
         await performBackmerge(
             instance(mockedGit),
@@ -705,7 +706,7 @@ describe("perform-backmerge with deprecated branches setting", () => {
         verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
         verify(mockedGit.checkout('master')).once();
         verify(mockedGit.configFetchAllRemotes()).once();
-        verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+        verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
 
         verify(mockedGit.checkout('develop')).once();
         verify(mockedGit.merge('master', 'theirs')).once();
@@ -725,7 +726,7 @@ describe("perform-backmerge with deprecated branches setting", () => {
         when(mockedGit.merge(anyString(), anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
 
         await performBackmerge(
             instance(mockedGit),
@@ -739,7 +740,7 @@ describe("perform-backmerge with deprecated branches setting", () => {
         verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
         verify(mockedGit.checkout('master')).once();
         verify(mockedGit.configFetchAllRemotes()).once();
-        verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+        verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
 
         verify(mockedGit.checkout('develop')).once();
         verify(mockedGit.merge('master', 'ours')).once();
@@ -904,12 +905,12 @@ describe("perform-backmerge to multiple branches with deprecated branches settin
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context;
         performBackmerge(instance(mockedGit), {branches: ['develop', 'dev']}, context)
             .then(() => {
                 verify(mockedGit.checkout('master')).twice();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.rebase('master')).twice();
 
                 verify(mockedLogger.error('Invalid branch configuration found and ignored.')).never();
@@ -936,7 +937,7 @@ describe("perform-backmerge to multiple branches with deprecated branches settin
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
         performBackmerge(instance(mockedGit), {branches: [{from: "main", to: "next"}]}, context)
             .then(() => {
                 verify(mockedLogger.log('Branch "next" was skipped as release did not originate from branch "main".')).once();
@@ -949,7 +950,7 @@ describe("perform-backmerge to multiple branches with deprecated branches settin
     it("abort when merging into the same branch and it's disallowed", async () => {
         const mockedGit = mock(Git);
         const mockedLogger = mock(NullLogger);
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
         await performBackmerge(instance(mockedGit), {branches: ['master', 'develop'], allowSameBranchMerge: false}, context);
         verify(mockedLogger.error('Process aborted due to an error while backmerging a branch.')).once();
         verify(mockedLogger.error(anyString())).once();
@@ -967,8 +968,12 @@ describe("perform-backmerge to multiple branches with deprecated branches settin
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
-        performBackmerge(instance(mockedGit), {branches: [{from: 'main', to: 'next'}, 'master'], allowSameBranchMerge: true}, context)
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
+        performBackmerge(
+            instance(mockedGit),
+            {branches: [{from: 'main', to: 'next'}, 'master'], allowSameBranchMerge: true},
+            context
+        )
             .then(() => {
                 verify(mockedLogger.log('Branch "next" was skipped as release did not originate from branch "main".')).once();
                 verify(mockedLogger.log('Performing back-merge into develop branch "next".')).never();
@@ -976,7 +981,7 @@ describe("perform-backmerge to multiple branches with deprecated branches settin
 
                 verify(mockedLogger.log('Performing back-merge into develop branch "master".')).once();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.checkout('master')).once();
                 verify(mockedGit.rebase('master')).never();
                 verify(mockedGit.push('my-repo', 'master', false)).once();
@@ -995,14 +1000,18 @@ describe("perform-backmerge to multiple branches with deprecated branches settin
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
-        performBackmerge(instance(mockedGit), {branches: [JSON.parse('{"foo": "bar"}'), 'master'], allowSameBranchMerge: true}, context)
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
+        performBackmerge(
+            instance(mockedGit),
+            {branches: [JSON.parse('{"foo": "bar"}'), 'master'], allowSameBranchMerge: true},
+            context
+        )
             .then(() => {
                 verify(mockedLogger.error('Invalid branch configuration found and ignored.')).once();
 
                 verify(mockedLogger.log('Performing back-merge into develop branch "master".')).once();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.checkout('master')).once();
                 verify(mockedGit.rebase('master')).never();
                 verify(mockedGit.push('my-repo', 'master', false)).once();
@@ -1021,12 +1030,16 @@ describe("perform-backmerge to multiple branches with deprecated branches settin
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
-        performBackmerge(instance(mockedGit), {branches: [{from: 'master', to: '${branch.name}'}], allowSameBranchMerge: true}, context)
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
+        performBackmerge(
+            instance(mockedGit),
+            {branches: [{from: 'master', to: '${branch.name}'}], allowSameBranchMerge: true},
+            context
+        )
             .then(() => {
                 verify(mockedLogger.log('Performing back-merge into develop branch "master".')).once();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.checkout('master')).once();
                 verify(mockedGit.rebase('master')).never();
                 verify(mockedGit.push('my-repo', 'master', false)).once();
@@ -1039,7 +1052,7 @@ describe("perform-backmerge to multiple branches with deprecated branches settin
 describe("perform-backmerge with error", () => {
     it("rebase with unstaged changes", (done) => {
         class MockError extends Error {
-            stderr: string
+            stderr= ''
         }
         const mockedGit = mock(Git);
         const mockedLogger = mock(NullLogger);
@@ -1050,14 +1063,14 @@ describe("perform-backmerge with error", () => {
         when(mockedGit.getModifiedFiles()).thenReturn(new Promise<string[]>(resolve => resolve(['M testfile'])));
         when(mockedLogger.error(anyString())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
         performBackmerge(instance(mockedGit), {branches: ['develop']}, context)
             .then(() => {
                 verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
                 verify(mockedLogger.error('Invalid branch configuration found and ignored.')).never();
                 verify(mockedGit.checkout('master')).once();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.checkout('develop')).once();
                 verify(mockedGit.rebase('master')).once();
                 verify(mockedGit.getModifiedFiles()).once();
@@ -1072,7 +1085,7 @@ describe("perform-backmerge with error", () => {
 
     it("rebase with other error", (done) => {
         class MockError extends Error {
-            stderr: string
+            stderr = ''
         }
         const mockedGit = mock(Git);
         const mockedLogger = mock(NullLogger);
@@ -1083,14 +1096,14 @@ describe("perform-backmerge with error", () => {
         when(mockedGit.rebase(anyString())).thenThrow({message:'',name:'',stderr:'any other error'} as MockError);
         when(mockedGit.getModifiedFiles()).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
         performBackmerge(instance(mockedGit), {branches: ['develop']}, context)
             .then(() => {
                 verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
                 verify(mockedLogger.error('Invalid branch configuration found and ignored.')).never();
                 verify(mockedGit.checkout('master')).once();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.checkout('develop')).once();
                 verify(mockedGit.rebase('master')).once();
                 verify(mockedGit.getModifiedFiles()).never();
@@ -1106,7 +1119,7 @@ describe("perform-backmerge with deprecated branchName setting", () => {
     it("trigger deprecation notice", async() => {
         const mockedGit = mock(Git);
         const mockedLogger = mock(NullLogger);
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context;
         await performBackmerge(instance(mockedGit), {branchName: 'master', allowSameBranchMerge: false}, context);
         verify(mockedLogger.log('The property "branchName" is deprecated. Please use "branches" instead: `branches: ["master"]`')).once();
         verify(mockedLogger.error(anyString())).once();
@@ -1122,13 +1135,13 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context;
         performBackmerge(instance(mockedGit), {branchName: 'develop'}, context)
             .then(() => {
                 verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
                 verify(mockedGit.checkout('master')).once();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.checkout('develop')).once();
                 verify(mockedGit.rebase('master')).once();
                 verify(mockedGit.push('my-repo', 'develop', false)).once();
@@ -1147,12 +1160,12 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
         performBackmerge(instance(mockedGit), {branchName: 'master', allowSameBranchMerge: true}, context)
             .then(() => {
                 verify(mockedLogger.log('Performing back-merge into develop branch "master".')).once();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.checkout('master')).once();
                 verify(mockedGit.rebase('master')).never();
                 verify(mockedGit.push('my-repo', 'master', false)).once();
@@ -1164,7 +1177,7 @@ describe("perform-backmerge with deprecated branchName setting", () => {
     it("disallow merging into the same branch", async () => {
         const mockedGit = mock(Git);
         const mockedLogger = mock(NullLogger);
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
         await performBackmerge(instance(mockedGit), {branchName: 'master', allowSameBranchMerge: false}, context);
         verify(mockedLogger.error(anyString())).once();
     });
@@ -1179,12 +1192,12 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context;
         performBackmerge(instance(mockedGit), {branchName: '${branch.name}', allowSameBranchMerge: true}, context)
             .then(() => {
                 verify(mockedLogger.log('Performing back-merge into develop branch "master".')).once();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.checkout('master')).once();
                 verify(mockedGit.rebase('master')).never();
                 verify(mockedGit.push('my-repo', 'master', false)).once();
@@ -1203,13 +1216,13 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
         performBackmerge(instance(mockedGit), {branchName: 'develop'}, context)
             .then(() => {
                 verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
                 verify(mockedGit.checkout('master')).once();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.checkout('develop')).once();
                 verify(mockedGit.rebase('master')).once();
                 verify(mockedGit.push('my-repo', 'develop', false)).once();
@@ -1228,13 +1241,13 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context;
         performBackmerge(instance(mockedGit), {branchName: 'develop', forcePush: true}, context)
             .then(() => {
                 verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
                 verify(mockedGit.checkout('master')).once();
                 verify(mockedGit.configFetchAllRemotes()).once();
-                verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+                verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
                 verify(mockedGit.checkout('develop')).once();
                 verify(mockedGit.rebase('master')).once();
                 verify(mockedGit.push('my-repo', 'develop', true)).once();
@@ -1255,7 +1268,7 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
 
         await performBackmerge(
             instance(mockedGit),
@@ -1268,7 +1281,7 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
         verify(mockedGit.checkout('master')).once();
         verify(mockedGit.configFetchAllRemotes()).once();
-        verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+        verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
         verify(mockedGit.checkout('develop')).once();
         verify(mockedGit.rebase('master')).once();
         verify(mockedGit.commit('my-commit-message')).once();
@@ -1287,7 +1300,7 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         when(mockedGit.rebase(anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
 
         await performBackmerge(
             instance(mockedGit),
@@ -1301,7 +1314,7 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
         verify(mockedGit.checkout('master')).once();
         verify(mockedGit.configFetchAllRemotes()).once();
-        verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+        verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
 
         const checkoutDevelopAction = mockedGit.checkout('develop');
         verify(mockedGit.stash()).calledBefore(checkoutDevelopAction);
@@ -1325,7 +1338,7 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         when(mockedGit.merge(anyString(), anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
 
         await performBackmerge(
             instance(mockedGit),
@@ -1338,7 +1351,7 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
         verify(mockedGit.checkout('master')).once();
         verify(mockedGit.configFetchAllRemotes()).once();
-        verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+        verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
 
         verify(mockedGit.checkout('develop')).once();
         verify(mockedGit.merge('master', 'none')).once();
@@ -1358,7 +1371,7 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         when(mockedGit.merge(anyString(), anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context;
 
         await performBackmerge(
             instance(mockedGit),
@@ -1372,7 +1385,7 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
         verify(mockedGit.checkout('master')).once();
         verify(mockedGit.configFetchAllRemotes()).once();
-        verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+        verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
 
         verify(mockedGit.checkout('develop')).once();
         verify(mockedGit.merge('master', 'ours')).once();
@@ -1392,7 +1405,7 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         when(mockedGit.merge(anyString(), anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
 
         await performBackmerge(
             instance(mockedGit),
@@ -1406,7 +1419,7 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
         verify(mockedGit.checkout('master')).once();
         verify(mockedGit.configFetchAllRemotes()).once();
-        verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+        verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
 
         verify(mockedGit.checkout('develop')).once();
         verify(mockedGit.merge('master', 'theirs')).once();
@@ -1426,7 +1439,7 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         when(mockedGit.merge(anyString(), anyString())).thenResolve();
         when(mockedGit.push(anyString(), anyString(), anything())).thenResolve();
 
-        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}};
+        const context = {logger: instance(mockedLogger), branch: {name: 'master'}, options: {repositoryUrl: 'my-repo'}} as Context
 
         await performBackmerge(
             instance(mockedGit),
@@ -1440,7 +1453,7 @@ describe("perform-backmerge with deprecated branchName setting", () => {
         verify(mockedLogger.log('Performing back-merge into develop branch "develop".')).once();
         verify(mockedGit.checkout('master')).once();
         verify(mockedGit.configFetchAllRemotes()).once();
-        verify(mockedGit.fetch(context.options.repositoryUrl)).once();
+        verify(mockedGit.fetch(context.options!.repositoryUrl)).once();
 
         verify(mockedGit.checkout('develop')).once();
         verify(mockedGit.merge('master', 'ours')).once();
